cretate README.MD
Self Schedule Discipliner is Portfolio Project at Hacktiv8 Bootcamp which was held since August 2016, and will be finish at this November 2016. I create this repository on my behalf to complete task and increase my project's numbers in github. Besides i need myself to be disciplined for tight time shecdule in this bootcamp.